Fix: Move DebugStat to be child of Viewport; fixes focus issue
Issue:
Before it was on a Separate Window positioned top-left of the Viewport. Clicking on anywhere on Main gives focus to the Main, hiding DebugStat behind the Main.

Solution:
Change DebugStat to be a Child Window of current Context rather than a separate Window.
And render DebugStat under Viewport.

Alternate solution:
Since there's no "always on top" or layer system, it can be solved by adding `ImGuiWindowFlags_NoBringToFrontOnFocus` to Main. This method can also be applied to any "layout region" type of Windows high on the hierarchy, but only because there's currently no Modal or window UI.