fix: Make `ErrorLog` a ChildWindow
It being Window allows other Window in Viewport can be brought to front
if that gains focus. When Viewport itself gets focused, since both have
takes the fullsize of vW & vH, it's un-switchable.

There was a bug where if a shader compile error occurs on startup,
it wasn't displayed. It's because the `ErrorLog` Window is first
created first, and then Viewport gets focused by `ui.Startup`.