refactor: Keep Renderer::texture as value
RATIONALE:
To make Renderer the owner of the texture it's producing.

Currently, output texture is not created deliberately, but it is just
taken from the last Node of the Compositor.

It will be created explicitly in the future, but the responsibility will
be determined once `OutputNode` is designed.

Until then, RenderToOutput() syncs with the Compositor's output texture
ID by just copying it each frame.